requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiohttp>=3.9.0
playwright>=1.50.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0
nest-asyncio>=1.6.0
google-generativeai>=0.8.0
pytest>=8.0.0
